backend/requirements.txt
fastapi
pymongo
requests
uvicorn
python-dotenv